move useGameViewportScaling to features/app-state/actions
