+++ src/features/app-state/reducer.js
@@ -5,18 +5,18 @@ const initialState = {
   sideMenu: false
 };
 
-const appStateReducer = (state = initialState, action) => {
+const appStateReducer = (state = initialState, { payload, type }) => {
 
-  switch(action.type) {
+  switch(type) {
 
     case 'OPT_OUT_DOWNLOAD':
       return { ...state, optOutDownload: true };
 
     case 'SET_LARGE_VIEW':
-      return { ...state, largeView: action.payload.value };
+      return { ...state, largeView: payload };
 
     case 'SET_SIDE_MENU':
-      return { ...state, sideMenu: action.payload.value };
+      return { ...state, sideMenu: payload };
 
     default:
       return state;
